requirements.txt
# Audio processing (cross-platform)
sounddevice>=0.4.6
numpy>=1.21.0
scipy>=1.10.0

# Global shortcuts and text injection (Linux-specific)
evdev>=1.6.0; sys_platform == 'linux'
pyperclip>=1.8.2

# macOS-specific dependencies
pyobjc-core>=9.0; sys_platform == 'darwin'
pyobjc-framework-Cocoa>=9.0; sys_platform == 'darwin'
pyobjc-framework-Quartz>=9.0; sys_platform == 'darwin'

# Whisper integration
pywhispercpp==1.3.3
requests>=2.25.0

# System integration
psutil>=5.8.0

# CLI formatting and logging
rich>=13.0.0
